JAV-469 move InstanceCacheManagerOld related functions into InstanceCacheManagerOld
