+++ service-registry/src/main/java/io/servicecomb/serviceregistry/registry/RemoteServiceRegistry.java
@@ -31,7 +31,6 @@ import io.servicecomb.serviceregistry.client.ServiceRegistryClient;
 import io.servicecomb.serviceregistry.client.http.ServiceRegistryClientImpl;
 import io.servicecomb.serviceregistry.config.ServiceRegistryConfig;
 import io.servicecomb.serviceregistry.definition.MicroserviceDefinition;
-import io.servicecomb.serviceregistry.task.InstancePullTask;
 import io.servicecomb.serviceregistry.task.event.PeriodicPullEvent;
 import io.servicecomb.serviceregistry.task.event.PullMicroserviceVersionsInstancesEvent;
 import io.servicecomb.serviceregistry.task.event.ShutdownEvent;
@@ -40,8 +39,6 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
   private static final Logger LOGGER = LoggerFactory.getLogger(RemoteServiceRegistry.class);
 
   private ScheduledThreadPoolExecutor taskPool;
-
-  private InstancePullTask pullTask;
 
   public RemoteServiceRegistry(EventBus eventBus, ServiceRegistryConfig serviceRegistryConfig,
       MicroserviceDefinition microserviceDefinition) {
@@ -62,7 +59,6 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
         LOGGER.warn("Too many pending tasks, reject " + task.getClass().getName());
       }
     });
-    pullTask = new InstancePullTask(serviceRegistryConfig.getInstancePullInterval(), this.instanceCacheManager);
   }
 
   @Override
@@ -87,11 +83,8 @@ public class RemoteServiceRegistry extends AbstractServiceRegistry {
         serviceRegistryConfig.getHeartbeatInterval(),
         TimeUnit.SECONDS);
     if (isNeedPull()) {
-      taskPool.scheduleAtFixedRate(pullTask,
-          serviceRegistryConfig.getInstancePullInterval(),
-          serviceRegistryConfig.getInstancePullInterval(),
-          TimeUnit.SECONDS);
-      taskPool.scheduleAtFixedRate(() -> eventBus.post(new PeriodicPullEvent()),
+      taskPool.scheduleAtFixedRate(
+          () -> eventBus.post(new PeriodicPullEvent()),
           serviceRegistryConfig.getInstancePullInterval(),
           serviceRegistryConfig.getInstancePullInterval(),
           TimeUnit.SECONDS);
